requirements.txt
tensorflow-gpu==1.12
keras==2.2.4
pillow==5.4.1
tqdm==4.31.1
pandas==0.24.1
matplotlib==3.0.2
